fix: 💫 update frame timing
