+++ CMakeLists.txt
@@ -45,9 +45,29 @@ target_link_libraries(
     sfml-system
     arkanoid_compiler_flags)
 
+add_library(Game_lib STATIC "${PROJECT_SOURCE_DIR}/src/Game.cpp")
+target_link_libraries(
+    Game_lib
+    sfml-network
+    sfml-audio
+    sfml-graphics
+    sfml-window
+    sfml-system
+    arkanoid_compiler_flags)
+
 add_library(Paddle_lib STATIC "${PROJECT_SOURCE_DIR}/src/Paddle.cpp")
 target_link_libraries(
     Paddle_lib
+    sfml-network
+    sfml-audio
+    sfml-graphics
+    sfml-window
+    sfml-system
+    arkanoid_compiler_flags)
+
+add_library(Rectangle_lib STATIC "${PROJECT_SOURCE_DIR}/src/Rectangle.cpp")
+target_link_libraries(
+    Rectangle_lib
     sfml-network
     sfml-audio
     sfml-graphics
@@ -60,7 +80,9 @@ target_link_libraries(
     Arkanoid
     PUBLIC Ball_lib
            Brick_lib
+           Game_lib
            Paddle_lib
+           Rectangle_lib
            arkanoid_compiler_flags)
 target_include_directories(Arkanoid PUBLIC "${PROJECT_BINARY_DIR}")
