+++ Catch_tests/CMakeLists.txt
@@ -15,6 +15,7 @@ target_link_libraries(
     PRIVATE Ball_lib
             Brick_lib
             Paddle_lib
+            Rectangle_lib
             arkanoid_compiler_flags)
 target_link_libraries(Catch_tests_run PRIVATE Catch2::Catch2WithMain)
 target_include_directories(Catch_tests_run PUBLIC "${PROJECT_SOURCE_DIR}/src")
