+++ src/Paddle.h
@@ -1,16 +1,12 @@
+#include "Rectangle.h"
+#include "constants.h"
+
 #include <SFML/Graphics.hpp>
 
-struct Paddle
+struct Paddle : public Rectangle
 {
-    Paddle(float initial_x, float initial_y);
-    void update();
-    float x() const;
-    float y() const;
-    float left() const;
-    float right() const;
-    float top() const;
-    float bottom() const;
-
-    sf::RectangleShape shape;
     sf::Vector2f velocity;
+
+    Paddle(float initial_x, float initial_y);
+    void update(FrameTime frame_time);
 };
