+++ src/constants.h
@@ -1,14 +1,18 @@
 #ifndef CONSTANTS_H
 #define CONSTANTS_H
 
+using FrameTime = float;
+
 namespace constants
 {
 inline constexpr unsigned int kWindowWidth{800U};
 inline constexpr unsigned int kWindowHeight{600U};
-inline constexpr unsigned int kFramerateLimit{60U};
+inline constexpr unsigned int kFramerateLimit{240U};
+inline constexpr float kFrameTimestep{1.F};
+inline constexpr float kFrameTimeSlice{1.F};
 
 inline constexpr float kBallRadius{10.F};
-inline constexpr float kBallVelocity{8.F};
+inline constexpr float kBallVelocity{0.8F};
 
 inline constexpr float kBlockWidth{60.F};
 inline constexpr float kBlockHeight{20.F};
@@ -18,7 +22,7 @@ inline constexpr float kBlocksInsetX{22.F};
 
 inline constexpr float kPaddleWidth{60.F};
 inline constexpr float kPaddleHeight{20.F};
-inline constexpr float kPaddleVelocity{6.F};
+inline constexpr float kPaddleVelocity{0.6F};
 inline constexpr float kPaddleInsetBottom{50.F};
 } // namespace constants
